% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{check_file_needs_processing}
\alias{check_file_needs_processing}
\title{Check if a file has already been processed by comparing modification times}
\usage{
check_file_needs_processing(conn, file_id, file_name, modified_time)
}
\arguments{
\item{conn}{Database connection}

\item{file_id}{Google Drive file ID}

\item{file_name}{File name}

\item{modified_time}{File modification time}
}
\value{
TRUE if file needs processing, FALSE if already processed and up to date
}
\description{
Check if a file has already been processed by comparing modification times
}
