% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_processing.R
\name{cleanup_temp_files}
\alias{cleanup_temp_files}
\title{Clean up temporary files}
\usage{
cleanup_temp_files(temp_dir)
}
\arguments{
\item{temp_dir}{Temporary directory to clean}
}
\value{
TRUE if successful
}
\description{
Clean up temporary files
}
