% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_aggregated_views}
\alias{create_aggregated_views}
\title{Create aggregated views for efficient querying}
\usage{
create_aggregated_views(conn)
}
\arguments{
\item{conn}{Database connection}
}
\value{
TRUE if successful
}
\description{
Create aggregated views for efficient querying
}
