% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_platform_metadata_table}
\alias{create_platform_metadata_table}
\title{Create the platform_metadata table}
\usage{
create_platform_metadata_table(conn, db_type = "sqlite")
}
\arguments{
\item{conn}{Database connection}

\item{db_type}{Database type}
}
\value{
TRUE if successful
}
\description{
Create the platform_metadata table
}
