% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_processed_files_table}
\alias{create_processed_files_table}
\title{Create the processed_files table}
\usage{
create_processed_files_table(conn, db_type = "sqlite")
}
\arguments{
\item{conn}{Database connection}

\item{db_type}{Database type}
}
\value{
TRUE if successful
}
\description{
Create the processed_files table
}
