% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{define_platforms}
\alias{define_platforms}
\title{Define music streaming platforms}
\usage{
define_platforms()
}
\value{
List of platform configurations
}
\description{
Define music streaming platforms
}
