% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{initialize_database_tables}
\alias{initialize_database_tables}
\title{Create necessary database tables}
\usage{
initialize_database_tables(conn, db_type = "sqlite")
}
\arguments{
\item{conn}{Database connection}

\item{db_type}{Database type}
}
\value{
TRUE if successful
}
\description{
Create necessary database tables
}
