% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{initialize_platforms}
\alias{initialize_platforms}
\title{Initialize all platforms in the database}
\usage{
initialize_platforms(conn)
}
\arguments{
\item{conn}{Database connection}
}
\value{
TRUE if successful
}
\description{
Initialize all platforms in the database
}
