% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{insert_platform_data}
\alias{insert_platform_data}
\title{Insert data into platform table}
\usage{
insert_platform_data(conn, df, platform_id)
}
\arguments{
\item{conn}{Database connection}

\item{df}{Data frame to insert}

\item{platform_id}{Platform ID}
}
\value{
Number of rows inserted
}
\description{
Insert data into platform table
}
