% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_processing.R
\name{preprocess_file}
\alias{preprocess_file}
\title{Process a file based on its type and extract if necessary}
\usage{
preprocess_file(file_path, platform, temp_dir = tempdir())
}
\arguments{
\item{file_path}{Path to the file}

\item{platform}{Platform configuration}

\item{temp_dir}{Temporary directory for extraction}
}
\value{
Path to the processed file or vector of paths if multiple files extracted
}
\description{
Process a file based on its type and extract if necessary
}
