% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_all_dsp_data}
\alias{process_all_dsp_data}
\title{Process all DSP data from Google Drive}
\usage{
process_all_dsp_data(
  system,
  parent_folder_id,
  date_pattern = "^[0-9]{8}$",
  date_range = NULL,
  temp_dir = tempdir()
)
}
\arguments{
\item{system}{System object from initialize_system}

\item{parent_folder_id}{Google Drive parent folder ID}

\item{date_pattern}{Regex pattern to identify date folders (default: "^[0-9]{8}$")}

\item{date_range}{Optional vector of length 2 with start and end dates}

\item{temp_dir}{Temporary directory for downloads (default: tempdir())}
}
\value{
Total number of files processed
}
\description{
Process all DSP data from Google Drive
}
