% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_drive.R
\name{process_gdrive_folder}
\alias{process_gdrive_folder}
\title{Process all new or modified files in a single folder}
\usage{
process_gdrive_folder(
  conn,
  folder_id,
  platforms = NULL,
  temp_dir = tempdir(),
  process_function = process_platform_file
)
}
\arguments{
\item{conn}{Database connection}

\item{folder_id}{Google Drive folder ID}

\item{platforms}{List of platform configurations}

\item{temp_dir}{Temporary directory for downloads}

\item{process_function}{Function to process each file}
}
\value{
Number of files processed
}
\description{
Process all new or modified files in a single folder
}
