% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_processing.R
\name{read_data_file}
\alias{read_data_file}
\title{Read a data file based on its type}
\usage{
read_data_file(file_path, platform)
}
\arguments{
\item{file_path}{Path to the file}

\item{platform}{Platform configuration}
}
\value{
Data frame containing the file contents
}
\description{
Read a data file based on its type
}
