% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{register_platform}
\alias{register_platform}
\title{Register a new platform in metadata}
\usage{
register_platform(conn, platform_id, platform_name, value_type = "unknown")
}
\arguments{
\item{conn}{Database connection}

\item{platform_id}{Platform ID}

\item{platform_name}{Platform name}

\item{value_type}{Type of value (streams, plays, views, purchases, etc.)}
}
\value{
TRUE if successful
}
\description{
Register a new platform in metadata
}
