% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{update_platform_metadata}
\alias{update_platform_metadata}
\title{Update platform metadata after processing files}
\usage{
update_platform_metadata(conn, platform_id, file_count = 1, record_count = 0)
}
\arguments{
\item{conn}{Database connection}

\item{platform_id}{Platform ID}

\item{file_count}{Number of files processed}

\item{record_count}{Number of records added}
}
\value{
TRUE if successful
}
\description{
Update platform metadata after processing files
}
